Extract CheckInsController
