+++ app/views/admin/reservations/check_ins/_form.html.erb
@@ -1,5 +1,5 @@
 <%= turbo_frame_tag "reservation-loan-form" do %>
-  <%= form_with(model: reservation_loan_lookup_form, url: check_in_admin_reservation_loans_path(reservation.id), builder: SpectreFormBuilder) do |form| %>
+  <%= form_with(model: reservation_loan_lookup_form, url: admin_reservation_check_ins_path(reservation.id), builder: SpectreFormBuilder) do |form| %>
     <%= form.text_field :reservable_item_id, label: "Item ID", autocomplete: "off", autofocus: true %>
     <%= form.submit "Return Item" %>
   <% end %>
