+++ app/views/admin/reservations/reservation_loans/index.html.erb
@@ -9,7 +9,7 @@
   <% end %>
 
   <% if @reservation.manager.state == :borrowed %>
-    <%= render partial: "admin/reservations/reservation_loans/lookup_form", locals: {reservation: @reservation, reservation_loan_lookup_form: ReservationLoanLookupForm.new} %>
+    <%= render partial: "admin/reservations/check_ins/form", locals: {reservation: @reservation, reservation_loan_lookup_form: ReservationLoanLookupForm.new} %>
   <% end %>
   <br>
